// FilmLibraryTest.Web/ClientApp/src/apiClient/apiClient.ts
export interface RegisterViewModel {
    email?: string | undefined;
    password?: string | undefined;
}

export interface IResultModelOfUserLoginViewModel {
    error?: string | undefined;
    result?: UserLoginViewModel | undefined;
}

export interface UserLoginViewModel {
    name?: string | undefined;
    token?: string | undefined;
}

export interface LoginViewModel {
    email?: string | undefined;
    password?: string | undefined;
}

export interface UserToTokenViewModel {
    token?: string | undefined;
}